// generated-typescript/models/product-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PC E-commerce API
 * Backend API for PC E-commerce Website - Nhom02 Website Ban PC
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: nhom02@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ProductDTO {
    'id'?: number;
    'name': string;
    'description'?: string;
    'price': number;
    'stock': number;
    'imageUrl'?: string;
    'isActive'?: boolean;
    'categoryId': number;
}